feat (panel): hook up more panel settings
